Added Euskara most common 1000 words
Description
Most frequent 1000 words:

front: word in spanish
back: word in euskara
The frequency of the words was determined using english vocabulary, due to lack of available analysis in Euskara for that matter. This means that a word may be very frequent in english but not so much in Euskara, and therefore this is just an approximation. I used the following web page for selecting the words, in descending order of frequency: https://strommeninc.com/1000-most-common-words-in-english-strommen-languages/